changed Image model to GroupImage and updated all corresponding migrations, seeders, routes, etc
